+++ src/cata_utility.h
@@ -10,8 +10,6 @@
 #include <utility>
 #include <vector>
 
-#include "units.h"
-
 class JsonIn;
 class JsonOut;
 class translation;
@@ -74,15 +72,6 @@ T divide_round_up( T num, T den )
     return ( num + den - 1 ) / den;
 }
 
-/** Divide @p num by @p den, rounding up
- *
- * @p num must be non-negative, @p den must be positive, and @c num+den must not overflow.
- */
-template<typename T, typename U>
-T divide_round_up( units::quantity<T, U> num, units::quantity<T, U> den )
-{
-    return divide_round_up( num.value(), den.value() );
-}
 
 int divide_round_down( int a, int b );
 
@@ -185,36 +174,6 @@ int bound_mod_to_vals( int val, int mod, int max, int min );
 const char *velocity_units( units_type vel_units );
 
 /**
- * Create a units label for a weight value.
- *
- * Gives the name of the weight unit in the user selected unit system, either
- * "kgs" or "lbs".  Used to add unit labels to the output of @ref convert_weight.
- *
- * @return name of unit
- */
-const char *weight_units();
-
-/**
- * Create an abbreviated units label for a volume value.
- *
- * Returns the abbreviated name for the volume unit for the user selected unit system,
- * i.e. "c", "L", or "qt". Used to add unit labels to the output of @ref convert_volume.
- *
- * @return name of unit.
- */
-const char *volume_units_abbr();
-
-/**
- * Create a units label for a volume value.
- *
- * Returns the abbreviated name for the volume unit for the user selected unit system,
- * ie "cup", "liter", or "quart". Used to add unit labels to the output of @ref convert_volume.
- *
- * @return name of unit.
- */
-const char *volume_units_long();
-
-/**
  * Convert internal velocity units to units defined by user.
  *
  * @param velocity A velocity value in internal units.
@@ -224,41 +183,6 @@ const char *volume_units_long();
  *          units for the object being measured.
  */
 double convert_velocity( int velocity, units_type vel_units );
-
-/**
- * Convert weight in grams to units defined by user (kg or lbs)
- *
- * @param weight to be converted.
- *
- * @returns Weight converted to user selected unit
- */
-double convert_weight( const units::mass &weight );
-
-/**
- * converts length to largest unit available
- * 1000 mm = 1 meter for example
- * assumed to be used in conjunction with unit string functions
- * also works for imperial units
- */
-int convert_length( const units::length &length );
-std::string length_units( const units::length &length );
-
-/** convert a mass unit to a string readable by a human */
-std::string weight_to_string( const units::mass &weight );
-
-/**
- * Convert volume from ml to units defined by user.
- */
-double convert_volume( int volume );
-
-/**
- * Convert volume from ml to units defined by user,
- * optionally returning the units preferred scale.
- */
-double convert_volume( int volume, int *out_scale );
-
-/** convert a volume unit to a string readable by a human */
-std::string vol_to_string( const units::volume &vol );
 
 /**
  * Convert a temperature from degrees Fahrenheit to degrees Celsius.
