+++ src/iuse.cpp
@@ -100,6 +100,7 @@
 #include "trap.h"
 #include "type_id.h"
 #include "ui.h"
+#include "units_utility.h"
 #include "value_ptr.h"
 #include "veh_type.h"
 #include "vehicle.h"
