+++ src/npctrade.cpp
@@ -35,6 +35,7 @@
 #include "type_id.h"
 #include "ui_manager.h"
 #include "units.h"
+#include "units_utility.h"
 #include "vehicle_selector.h"
 #include "visitable.h"
