+++ src/point.cpp
@@ -5,6 +5,7 @@
 #include <utility>
 
 #include "cata_utility.h"
+#include "units_utility.h"
 
 point point::from_string( const std::string &s )
 {
