+++ src/veh_interact.cpp
@@ -56,6 +56,7 @@
 #include "ui.h"
 #include "ui_manager.h"
 #include "units.h"
+#include "units_utility.h"
 #include "value_ptr.h"
 #include "veh_type.h"
 #include "veh_utils.h"
