+++ src/vehicle_display.cpp
@@ -8,9 +8,11 @@
 #include "calendar.h"
 #include "cata_utility.h"
 #include "catacharset.h"
+#include "color.h"
 #include "cursesdef.h"
 #include "debug.h"
 #include "itype.h"
+#include "optional.h"
 #include "options.h"
 #include "output.h"
 #include "string_formatter.h"
@@ -18,8 +20,7 @@
 #include "veh_type.h"
 #include "vpart_position.h"
 #include "units.h"
-#include "color.h"
-#include "optional.h"
+#include "units_utility.h"
 
 static const std::string part_location_structure( "structure" );
 static const itype_id itype_battery( "battery" );
